* Rewrote GiveMoney function and created several offspring Give*Money functions * Added GroupPowers enum courtesy of Thoys and switched GroupManager over to use it * Renamed Simulator.SimCaps to Simulator.Caps * Turned off Keep-Alive and Expect-100: Continue headers on Login * Added MoneyTransactionType and TransactionFlags
git-svn-id: http://libopenmetaverse.googlecode.com/svn/trunk@1391 52acb1d6-8a22-11de-b505-999d5b087335

+++ libsecondlife/examples/groupmanager/frmGroupInfo.cs
@@ -218,7 +218,7 @@ namespace groupmanager
                             memberData.ID = member.ID;
                             memberData.IsOwner = member.IsOwner;
                             memberData.LastOnline = member.OnlineStatus;
-                            memberData.Powers = member.Powers;
+                            memberData.Powers = (ulong)member.Powers;
                             memberData.Title = member.Title;
                             memberData.Contribution = member.Contribution;
